Descend from max instead of ascend from min
Previously to access versions which were not the minimum or maximum item in the list, required starting from the minimum version, and ascending through the versions of the list until it found the relevant version.

Now, it starts at the maximum item and descends through the versions of the list until it finds the relevant version.

This ensures that writes where the version number always increments (as is probably the case in most situations), are more efficient, as writes are 0(1) as opposed to O(n).